Insurance-Claim-Project/backend/requirements.txt
psycopg2
Django
djangorestframework
scikit-learn
tensorflow
joblib
django-cors-headers
djangorestframework-simplejwt
djoser
django-environ
